Move crawl-related models to models module
